update setup, add commands module, reorganize main
