+++ vodbot/__init__.py
@@ -1,2 +1,2 @@
 __project__ = "VodBot"
-__version__ = "1.2.0"
+__version__ = "1.3.0"
