+++ vodbot/commands/init.py
@@ -1,0 +1,4 @@
+# Initialization of VodBot, setting up configs, etc.
+
+def run(args):
+	pass
