+++ vodbot/commands/stage.py
@@ -1,0 +1,9 @@
+# Staging, where videos get staged and set up with metadata to upload
+
+def run(args):
+	if args.action == "add":
+		pass
+	elif args.action == "rm":
+		pass
+	elif args.action == "list":
+		pass
