+++ vodbot/commands/upload.py
@@ -1,0 +1,4 @@
+# Upload, command to upload staged videos to YouTube
+
+def run(args):
+	pass
