+++ vodbot/util.py
@@ -1,3 +1,5 @@
+# Module to pull and create different files and directories on the OS
+
 from .printer import cprint
 
 import os
